Added Render.quality and Render.resetQuality()
Set Render.quality to temporarily override the quality of the stage.
Call Render.resetQuality() to reset it back to the default stage
quality. You can pass a StageQuality value to resetQuality() to
change the default quality.

Render.highQuality has been renamed to Render.graphicsQuality, and
represents the quality used when Render is drawing vectors.